Imagemagick: patch for xml2-config
Imagemagick-6.7.x uses libxml2, but Apple changed the file
`xml2-config` starting in Mt. Lion, removing the `--prefix` flag.
That breaks Imagemagick, which uses it to find the library.
Because all versions of OSX have libxml2 in the same location,
this commit is safe for all users.

* Patch `configure` to not run `xml2-config --prefix` by placing
the value it returns, `/usr`, into the corresonding variable.

Fixes #11789.
Closes #11883.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>